stanford/LAVIS/requirements.txt
contexttimer
decord
diffusers<=0.16.0
einops>=0.4.1
fairscale==0.4.4
ftfy
iopath
ipython
omegaconf
opencv-python-headless==4.5.5.64
opendatasets
packaging
pandas
plotly
pre-commit
pycocoevalcap
pycocotools
python-magic
scikit-image
sentencepiece
spacy
streamlit
timm==0.4.12
torch==2.0.1
torchvision
tqdm
transformers==4.25.0
webdataset
wheel
tensorboard
install-jdk
cycling_utils @ git+https://github.com/StrongResearch/cycling_utils.git

